varios cambios

+++ src/main/java/com/franco/dev/graphql/financiero/GastoGraphQL.java
@@ -12,6 +12,7 @@ import com.franco.dev.service.financiero.TipoGastoService;
 import com.franco.dev.service.impresion.ImpresionService;
 import com.franco.dev.service.personas.FuncionarioService;
 import com.franco.dev.service.personas.UsuarioService;
+import com.franco.dev.utilitarios.StringUtils;
 import graphql.GraphQLException;
 import graphql.kickstart.tools.GraphQLMutationResolver;
 import graphql.kickstart.tools.GraphQLQueryResolver;
@@ -51,34 +52,36 @@ public class GastoGraphQL implements GraphQLQueryResolver, GraphQLMutationResolv
     @Autowired
     private MultiTenantService multiTenantService;
 
-    public Optional<Gasto> gasto(Long id, Long sucId) {
-        return service.findById(new EmbebedPrimaryKey(id, sucId));
+    public Gasto gasto(Long id, Long sucId) {
+        return multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.findByIdAndSucursalId((Long) params[0], (Long) params[1]), id, sucId);
     }
 
     public List<Gasto> gastos(int page, int size, Long sucId) {
         Pageable pageable = PageRequest.of(page, size);
-        return service.findAll(pageable);
+        return multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.findAll(pageable), pageable);
     }
 
     public List<Gasto> gastosPorCajaId(Long id, Long sucId) {
-        return service.findByCajaId(id, sucId);
+        return multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.findByCajaId(id, sucId), id, sucId);
     }
 
     public List<Gasto> gastosPorFecha(String inicio, String fin, Long sucId) {
-        return service.findByDate(inicio, fin, sucId);
+        return multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.findByDate(inicio, fin, sucId), inicio, fin, sucId);
     }
 
     public Gasto saveGasto(GastoInput input, String printerName, String local) throws GraphQLException {
         ModelMapper m = new ModelMapper();
         Gasto e = m.map(input, Gasto.class);
-        e = service.save(e);
-        multiTenantService.compartir(null, (Gasto s) -> service.save(s), e);
+//        e = service.save(e);
+//        multiTenantService.compartir(null, (Gasto s) -> service.save(s), e);
         return e;
     }
 
-    public List<Gasto> filterGastos(Long id, Long cajaId, Long sucId, Long responsableId, Integer page, Integer size){
+    public List<Gasto> filterGastos(Long id, Long cajaId, Long sucId, Long responsableId, String descripcion, Integer page, Integer size){
         Pageable pageable = PageRequest.of(page, size);
-        return service.filterGastos(id, cajaId, sucId, responsableId, pageable);
+        descripcion = StringUtils.convertToCustomFormat(descripcion);
+        String finalDescripcion = descripcion;
+        return multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.filterGastos(id, cajaId, sucId, responsableId, finalDescripcion, pageable), id, cajaId, sucId, responsableId, finalDescripcion, pageable);
     }
 
 //    public List<Gasto> gastosSearch(String texto){
@@ -86,9 +89,12 @@ public class GastoGraphQL implements GraphQLQueryResolver, GraphQLMutationResolv
 //    }
 
     public Boolean deleteGasto(Long id, Long sucId) {
-        Gasto gasto = service.findById(new EmbebedPrimaryKey(id, sucId)).orElse(null);
-        multiTenantService.compartir("filial"+sucId+"_bkp", (Gasto s) -> service.delete(s), gasto);
-        return service.delete(gasto);
+        Gasto gasto = multiTenantService.compartir("filial"+sucId+"_bkp", (params) -> service.findByIdAndSucursalId(id, sucId), id, sucId);
+        if (gasto != null) {
+            return multiTenantService.compartir("filial"+sucId+"_bkp", (Gasto s) -> service.delete(s), gasto);
+        } else {
+            throw new GraphQLException("No se pudo eliminar el gasto");
+        }
     }
 
     public Long countGasto() {
